assignment_2/requirements.txt
tensorflow_cpu==2.2.0
joblib==0.14.1
matplotlib==3.1.3
numpy==1.18.1
